getAssetModelsWithItems

+++ Backend/controllers/assetModelController.js
@@ -1,3 +1,4 @@
+import AssetItem from "../models/AssetItem.js";
 import AssetModel from "../models/AssetModel.js";
 
 export const addAsset = async (req, res) => {
@@ -48,17 +49,55 @@ export const editAssetModel = async (req, res) => {
 };
 
 
-export const deleteAssetModel = async (req, res) => {
+export const deleteAssetModel = async (req,res) => {
     try {
-        let { id } = req.query  //get the id from req query ?id=
-        let response = await AssetModel.findByIdAndDelete(id)
-        if (response) {
-            return res.status(200).json({ message: "Asset Model Deleted" })
-        } else {
-            return res.status(400).json({ message: "Asset Model Not Found" })
+        let { id } = req.query  //get the id from req query ?id=123456789
+        let isModel = await AssetModel.findById(id)
+        if (isModel) {
+
+            let isItems = await AssetItem.find({model: id})
+            if (isItems) {
+                return res.status(400).json({ message: "can't delete the model used for items" })
+            }else{
+                 await AssetModel.findByIdAndDelete(id)
+            return res.status(200).json({ message: "Assetmodel deleted" })
+            }
+           
         }
 
     } catch (error) {
         return res.status(500).json({ message: "Server error", error: error.message });
     }
 };
+
+
+export const getAssetModel = async (req, res) => {
+    try {
+        const response = await AssetModel.find()
+
+        return res.status(200).send(response)
+    } catch (error) {
+        return res.status(500).send({ message: "Something went wrong", error: error.message })
+    }
+}
+
+
+export const getAssetModelsWithItems = async (req, res) => {
+    try {
+
+        let allModelsWithItems = await AssetModel.aggregate([{
+            $lookup: {
+                localField: "_id",
+                from: "assetitems",
+                foreignField: 'model',
+                as: "items",
+            }
+        }])
+        res.status(200).send(allModelsWithItems)
+
+    } catch (error) {
+        return res.status(500).send({ message: "Something went wrong", error: error.message })
+    }
+}
+
+
